FAC-1009:impl:Rename translate package to translations
