+++ flickit-assessment-kit/src/test/java/org/flickit/assessment/kit/application/port/in/attribute/UpdateAttributeUseCaseParamTest.java
@@ -2,7 +2,7 @@ package org.flickit.assessment.kit.application.port.in.attribute;
 
 import jakarta.validation.ConstraintViolationException;
 import org.apache.commons.lang3.RandomStringUtils;
-import org.flickit.assessment.common.application.domain.kit.translate.AttributeTranslation;
+import org.flickit.assessment.common.application.domain.kit.translation.AttributeTranslation;
 import org.flickit.assessment.common.exception.ValidationException;
 import org.flickit.assessment.kit.application.port.in.attribute.UpdateAttributeUseCase.Param;
 import org.junit.jupiter.api.Test;
